requirements.txt
Django==2.1.2
pytest==3.8.1
djangorestframework==3.8.2
drf-yasg=1.10.1
model-mommy==1.6.0
django-rest-swagger==2.2.0
django-filter==2.0.0
